Refactor analyze module into submodules
