+++ tailtriage-cli/src/analyze/mod.rs
@@ -3,6 +3,23 @@ use std::collections::{BTreeMap, HashMap};
 use serde::{Serialize, Serializer};
 use tailtriage_core::{InFlightSnapshot, Run, RuntimeSnapshot};
 
+mod confidence;
+mod evidence;
+mod render;
+mod route;
+mod scoring;
+mod temporal;
+
+use confidence::apply_evidence_aware_confidence_caps;
+use evidence::{evidence_quality, truncation_warnings};
+pub use render::render_text;
+use route::route_breakdowns;
+use scoring::{
+    blocking_pressure_suspect, downstream_stage_suspect, executor_pressure_suspect,
+    queue_saturation_suspect,
+};
+use temporal::temporal_segments;
+
 const LOW_COMPLETED_REQUEST_THRESHOLD: usize = 20;
 const QUEUE_SHARE_TRIGGER_PERMILLE: u64 = 300;
 const MEDIUM_CONFIDENCE_SCORE_THRESHOLD: u8 = 65;
@@ -452,993 +469,6 @@ fn analyze_run_internal(run: &Run) -> Report {
     }
 }
 
-struct RouteBreakdownContext {
-    breakdowns: Vec<RouteBreakdown>,
-    divergent: bool,
-}
-
-fn route_breakdowns(run: &Run, global: &Report) -> RouteBreakdownContext {
-    let mut ids_by_route: BTreeMap<String, Vec<String>> = BTreeMap::new();
-    for request in &run.requests {
-        ids_by_route
-            .entry(request.route.clone())
-            .or_default()
-            .push(request.request_id.clone());
-    }
-    let eligible: Vec<(String, Vec<String>)> = ids_by_route
-        .into_iter()
-        .filter(|(_, ids)| ids.len() >= ROUTE_MIN_REQUEST_COUNT)
-        .collect();
-    if eligible.len() < 2 {
-        return RouteBreakdownContext {
-            breakdowns: vec![],
-            divergent: false,
-        };
-    }
-
-    let omitted_routes = run
-        .requests
-        .iter()
-        .fold(BTreeMap::<String, usize>::new(), |mut acc, request| {
-            *acc.entry(request.route.clone()).or_default() += 1;
-            acc
-        })
-        .into_values()
-        .filter(|count| *count < ROUTE_MIN_REQUEST_COUNT)
-        .count();
-
-    let mut candidates = Vec::new();
-    for (route, request_ids) in eligible {
-        let filtered = filtered_run_for_route(run, &request_ids);
-        let mut analyzed = analyze_run_internal(&filtered);
-        analyzed
-            .warnings
-            .push(ROUTE_RUNTIME_ATTRIBUTION_WARNING.to_string());
-        candidates.push(RouteBreakdown {
-            route,
-            request_count: analyzed.request_count,
-            p50_latency_us: analyzed.p50_latency_us,
-            p95_latency_us: analyzed.p95_latency_us,
-            p99_latency_us: analyzed.p99_latency_us,
-            p95_queue_share_permille: analyzed.p95_queue_share_permille,
-            p95_service_share_permille: analyzed.p95_service_share_permille,
-            evidence_quality: analyzed.evidence_quality,
-            primary_suspect: analyzed.primary_suspect,
-            secondary_suspects: analyzed.secondary_suspects,
-            warnings: analyzed.warnings,
-        });
-    }
-    if !should_emit_route_breakdowns(global, &candidates) {
-        return RouteBreakdownContext {
-            breakdowns: vec![],
-            divergent: false,
-        };
-    }
-    let mut emitted = candidates;
-    emitted.sort_by(|a, b| {
-        b.p95_latency_us
-            .cmp(&a.p95_latency_us)
-            .then_with(|| b.request_count.cmp(&a.request_count))
-            .then_with(|| a.route.cmp(&b.route))
-    });
-    emitted.truncate(ROUTE_BREAKDOWN_LIMIT);
-    let divergent = route_divergence(&emitted);
-    if omitted_routes > 0 {
-        let note = format!(
-            "Some routes are omitted from route_breakdowns because they have fewer than {ROUTE_MIN_REQUEST_COUNT} completed requests."
-        );
-        for breakdown in &mut emitted {
-            breakdown.warnings.push(note.clone());
-        }
-    }
-    RouteBreakdownContext {
-        breakdowns: emitted,
-        divergent,
-    }
-}
-
-fn route_divergence(candidates: &[RouteBreakdown]) -> bool {
-    candidates
-        .iter()
-        .map(|c| c.primary_suspect.kind.as_str())
-        .collect::<std::collections::BTreeSet<_>>()
-        .len()
-        >= 2
-}
-
-fn should_emit_route_breakdowns(global: &Report, candidates: &[RouteBreakdown]) -> bool {
-    if candidates.len() < 2 {
-        return false;
-    }
-    if route_divergence(candidates) {
-        return true;
-    }
-    let p95s: Vec<u64> = candidates.iter().filter_map(|c| c.p95_latency_us).collect();
-    if p95s.len() < 2 {
-        return false;
-    }
-    let slowest = *p95s.iter().max().unwrap_or(&0);
-    let fastest = *p95s.iter().min().unwrap_or(&0);
-    (fastest > 0 && slowest.saturating_mul(2) >= fastest.saturating_mul(3))
-        || match global.p95_latency_us {
-            Some(global_p95) if global_p95 > 0 => {
-                slowest.saturating_mul(4) >= global_p95.saturating_mul(5)
-            }
-            _ => false,
-        }
-}
-
-fn filtered_run_for_route(run: &Run, request_ids: &[String]) -> Run {
-    let request_ids: std::collections::HashSet<&str> =
-        request_ids.iter().map(String::as_str).collect();
-    let mut filtered = run.clone();
-    filtered.requests = run
-        .requests
-        .iter()
-        .filter(|r| request_ids.contains(r.request_id.as_str()))
-        .cloned()
-        .collect();
-    filtered.stages = run
-        .stages
-        .iter()
-        .filter(|s| request_ids.contains(s.request_id.as_str()))
-        .cloned()
-        .collect();
-    filtered.queues = run
-        .queues
-        .iter()
-        .filter(|q| request_ids.contains(q.request_id.as_str()))
-        .cloned()
-        .collect();
-    filtered.runtime_snapshots = Vec::new();
-    filtered.inflight = Vec::new();
-    filtered
-}
-
-fn filtered_run_for_temporal_segment(
-    run: &Run,
-    request_ids: &[String],
-    start: u64,
-    end: u64,
-) -> Run {
-    let mut filtered = filtered_run_for_route(run, request_ids);
-    filtered.runtime_snapshots = run
-        .runtime_snapshots
-        .iter()
-        .filter(|s| s.at_unix_ms >= start && s.at_unix_ms <= end)
-        .cloned()
-        .collect();
-    filtered.inflight = run
-        .inflight
-        .iter()
-        .filter(|s| s.at_unix_ms >= start && s.at_unix_ms <= end)
-        .cloned()
-        .collect();
-    filtered
-}
-
-fn temporal_segments(run: &Run, global_warnings: &mut Vec<String>) -> Vec<TemporalSegment> {
-    if run.requests.len() < TEMPORAL_MIN_REQUEST_COUNT {
-        return vec![];
-    }
-    let mut requests = run.requests.clone();
-    requests.sort_by(|a, b| {
-        a.started_at_unix_ms
-            .cmp(&b.started_at_unix_ms)
-            .then_with(|| a.request_id.cmp(&b.request_id))
-    });
-    let split = requests.len() / 2;
-    let (early, late) = requests.split_at(split);
-    if early.len() < TEMPORAL_MIN_SEGMENT_REQUEST_COUNT
-        || late.len() < TEMPORAL_MIN_SEGMENT_REQUEST_COUNT
-    {
-        return vec![];
-    }
-    let build = |name: &str, seg: &[tailtriage_core::RequestEvent]| {
-        let ids: Vec<String> = seg.iter().map(|r| r.request_id.clone()).collect();
-        let start = seg.iter().map(|r| r.started_at_unix_ms).min();
-        let finish = seg.iter().map(|r| r.finished_at_unix_ms).max();
-        let mut analyzed = match (start, finish) {
-            (Some(s), Some(f)) => {
-                analyze_run_internal(&filtered_run_for_temporal_segment(run, &ids, s, f))
-            }
-            _ => analyze_run_internal(&filtered_run_for_route(run, &ids)),
-        };
-        let sparse_runtime =
-            analyzed.evidence_quality.runtime_snapshots != SignalCoverageStatus::Present;
-        let sparse_inflight =
-            analyzed.evidence_quality.inflight_snapshots != SignalCoverageStatus::Present;
-        if matches!(
-            analyzed.primary_suspect.kind,
-            DiagnosisKind::ExecutorPressureSuspected | DiagnosisKind::BlockingPoolPressure
-        ) && (sparse_runtime || sparse_inflight)
-        {
-            analyzed
-                .warnings
-                .push(TEMPORAL_RUNTIME_ATTRIBUTION_WARNING.to_string());
-        }
-        TemporalSegment {
-            name: name.to_string(),
-            request_count: analyzed.request_count,
-            started_at_unix_ms: start,
-            finished_at_unix_ms: finish,
-            p50_latency_us: analyzed.p50_latency_us,
-            p95_latency_us: analyzed.p95_latency_us,
-            p99_latency_us: analyzed.p99_latency_us,
-            p95_queue_share_permille: analyzed.p95_queue_share_permille,
-            p95_service_share_permille: analyzed.p95_service_share_permille,
-            evidence_quality: analyzed.evidence_quality,
-            primary_suspect: analyzed.primary_suspect,
-            secondary_suspects: analyzed.secondary_suspects,
-            warnings: analyzed.warnings,
-        }
-    };
-    let mut early_seg = build("early", early);
-    let mut late_seg = build("late", late);
-    let suspect_shift_raw = early_seg.primary_suspect.kind != late_seg.primary_suspect.kind;
-    let p95_shift = has_material_p95_shift(early_seg.p95_latency_us, late_seg.p95_latency_us);
-    let queue_move = matches!((early_seg.p95_queue_share_permille, late_seg.p95_queue_share_permille), (Some(a), Some(b)) if a.abs_diff(b) >= TEMPORAL_SHARE_SHIFT_PERMILLE);
-    let service_move = matches!((early_seg.p95_service_share_permille, late_seg.p95_service_share_permille), (Some(a), Some(b)) if a.abs_diff(b) >= TEMPORAL_SHARE_SHIFT_PERMILLE);
-    let runtime_sparse = early_seg.evidence_quality.runtime_snapshots
-        != SignalCoverageStatus::Present
-        || early_seg.evidence_quality.inflight_snapshots != SignalCoverageStatus::Present
-        || late_seg.evidence_quality.runtime_snapshots != SignalCoverageStatus::Present
-        || late_seg.evidence_quality.inflight_snapshots != SignalCoverageStatus::Present;
-    let runtime_dependent_shift = matches!(
-        (
-            &early_seg.primary_suspect.kind,
-            &late_seg.primary_suspect.kind
-        ),
-        (
-            DiagnosisKind::ExecutorPressureSuspected | DiagnosisKind::BlockingPoolPressure,
-            _
-        ) | (
-            _,
-            DiagnosisKind::ExecutorPressureSuspected | DiagnosisKind::BlockingPoolPressure
-        )
-    );
-    let suspect_shift = suspect_shift_raw
-        && (!runtime_sparse || !runtime_dependent_shift || p95_shift || queue_move || service_move);
-    let material = suspect_shift || p95_shift || queue_move || service_move;
-    if !material {
-        return vec![];
-    }
-    if suspect_shift {
-        global_warnings.push(TEMPORAL_SUSPECT_SHIFT_WARNING.to_string());
-    }
-    if p95_shift {
-        global_warnings.push(TEMPORAL_P95_SHIFT_WARNING.to_string());
-    }
-    apply_temporal_overlap_attribution_warning(&mut early_seg, &mut late_seg);
-    vec![early_seg, late_seg]
-}
-
-fn apply_temporal_overlap_attribution_warning(
-    early_seg: &mut TemporalSegment,
-    late_seg: &mut TemporalSegment,
-) {
-    let windows_overlap = matches!(
-        (
-            early_seg.started_at_unix_ms,
-            early_seg.finished_at_unix_ms,
-            late_seg.started_at_unix_ms,
-            late_seg.finished_at_unix_ms,
-        ),
-        (Some(_), Some(early_finish), Some(late_start), Some(_)) if early_finish >= late_start
-    );
-    let has_segment_runtime_or_inflight_samples = early_seg.evidence_quality.runtime_snapshot_count
-        > 0
-        || early_seg.evidence_quality.inflight_snapshot_count > 0
-        || late_seg.evidence_quality.runtime_snapshot_count > 0
-        || late_seg.evidence_quality.inflight_snapshot_count > 0;
-    if windows_overlap && has_segment_runtime_or_inflight_samples {
-        early_seg
-            .warnings
-            .push(TEMPORAL_OVERLAP_ATTRIBUTION_WARNING.to_string());
-        late_seg
-            .warnings
-            .push(TEMPORAL_OVERLAP_ATTRIBUTION_WARNING.to_string());
-    }
-}
-
-fn has_material_p95_shift(left: Option<u64>, right: Option<u64>) -> bool {
-    let (Some(a), Some(b)) = (left, right) else {
-        return false;
-    };
-    let lower = a.min(b);
-    let higher = a.max(b);
-    if lower == 0 {
-        return false;
-    }
-    higher.saturating_mul(2) >= lower.saturating_mul(3)
-}
-
-fn apply_evidence_aware_confidence_caps(
-    suspects: &mut [Suspect],
-    run: &Run,
-    evidence_quality: &EvidenceQuality,
-) {
-    let runtime_snapshots_missing = run.runtime_snapshots.is_empty();
-    let runtime_partial_key_fields = !runtime_snapshots_missing
-        && (run
-            .runtime_snapshots
-            .iter()
-            .all(|snapshot| snapshot.blocking_queue_depth.is_none())
-            || run
-                .runtime_snapshots
-                .iter()
-                .all(|snapshot| snapshot.local_queue_depth.is_none())
-            || run
-                .runtime_snapshots
-                .iter()
-                .all(|snapshot| snapshot.global_queue_depth.is_none()));
-    let ambiguous_cluster = ambiguity_cluster_indices(suspects);
-    for (i, suspect) in suspects.iter_mut().enumerate() {
-        let mut cap = Confidence::High;
-        let mut notes = Vec::new();
-        let is_primary = i == 0;
-        let is_insufficient = suspect.kind == DiagnosisKind::InsufficientEvidence;
-        if !is_insufficient && evidence_quality.quality == EvidenceQualityLevel::Weak {
-            cap = cap.min(Confidence::Medium);
-        }
-        if !is_insufficient && run.requests.is_empty() {
-            cap = Confidence::Low;
-            notes.push("Low completed-request count caps confidence.".to_string());
-        } else if run.requests.len() < LOW_COMPLETED_REQUEST_THRESHOLD {
-            if !is_insufficient {
-                cap = cap.min(Confidence::Medium);
-            }
-            if is_primary {
-                notes.push("Low completed-request count caps confidence.".to_string());
-            }
-        }
-        if run.truncation.dropped_requests > 0 && !is_insufficient {
-            cap = cap.min(Confidence::Medium);
-            notes.push(
-                "Capture truncation caps confidence because dropped evidence may affect ranking."
-                    .to_string(),
-            );
-        }
-        apply_family_evidence_caps(
-            &suspect.kind,
-            run,
-            runtime_snapshots_missing,
-            runtime_partial_key_fields,
-            &mut cap,
-            &mut notes,
-        );
-        let ambiguity_capped = ambiguous_cluster.contains(&i) && !is_insufficient;
-        if ambiguity_capped {
-            cap = cap.min(Confidence::Medium);
-            notes.push(
-                "Top suspects are close in score; confidence is capped by ambiguity.".to_string(),
-            );
-        }
-        let original = suspect.confidence;
-        suspect.confidence = original.min(cap);
-        let cap_changed_bucket = suspect.confidence != original;
-        if cap_changed_bucket || ambiguity_capped {
-            notes.sort();
-            notes.dedup();
-            suspect.confidence_notes = notes;
-        } else {
-            suspect.confidence_notes.clear();
-        }
-    }
-}
-
-fn apply_family_evidence_caps(
-    kind: &DiagnosisKind,
-    run: &Run,
-    runtime_snapshots_missing: bool,
-    runtime_partial_key_fields: bool,
-    cap: &mut Confidence,
-    notes: &mut Vec<String>,
-) {
-    match kind {
-        DiagnosisKind::ApplicationQueueSaturation => {
-            if run.truncation.dropped_queues > 0 {
-                *cap = (*cap).min(Confidence::Medium);
-                notes.push(
-                    "Capture truncation caps confidence because dropped evidence may affect ranking."
-                        .to_string(),
-                );
-            }
-            if run.queues.is_empty() {
-                *cap = (*cap).min(Confidence::Medium);
-                notes.push(
-                    "Missing queue instrumentation limits queue-saturation confidence.".to_string(),
-                );
-            }
-        }
-        DiagnosisKind::DownstreamStageDominates => {
-            if run.truncation.dropped_stages > 0 {
-                *cap = (*cap).min(Confidence::Medium);
-                notes.push(
-                    "Capture truncation caps confidence because dropped evidence may affect ranking."
-                        .to_string(),
-                );
-            }
-            if run.stages.is_empty() {
-                *cap = (*cap).min(Confidence::Medium);
-                notes.push(
-                    "Missing stage instrumentation limits downstream-stage confidence.".to_string(),
-                );
-            }
-        }
-        DiagnosisKind::BlockingPoolPressure | DiagnosisKind::ExecutorPressureSuspected => {
-            if run.truncation.dropped_runtime_snapshots > 0 {
-                *cap = (*cap).min(Confidence::Medium);
-                notes.push(
-                    "Capture truncation caps confidence because dropped evidence may affect ranking."
-                        .to_string(),
-                );
-            }
-            if runtime_snapshots_missing {
-                *cap = (*cap).min(Confidence::Medium);
-                notes.push(
-                    "Missing runtime snapshots limit executor/blocking confidence.".to_string(),
-                );
-            } else if runtime_partial_key_fields {
-                *cap = (*cap).min(Confidence::Medium);
-                notes.push(
-                    "Runtime snapshots are partial; missing runtime queue-depth fields limit executor/blocking confidence.".to_string(),
-                );
-            }
-        }
-        DiagnosisKind::InsufficientEvidence => {}
-    }
-}
-
-fn ambiguity_cluster_indices(suspects: &[Suspect]) -> Vec<usize> {
-    let mut ranked = suspects
-        .iter()
-        .enumerate()
-        .filter(|(_, s)| s.kind != DiagnosisKind::InsufficientEvidence)
-        .collect::<Vec<_>>();
-    ranked.sort_by_key(|(_, s)| std::cmp::Reverse(s.score));
-    let Some((_, top)) = ranked.first() else {
-        return Vec::new();
-    };
-    if top.score < AMBIGUITY_MIN_SCORE_THRESHOLD {
-        return Vec::new();
-    }
-    let cluster = ranked
-        .iter()
-        .take_while(|(_, s)| {
-            s.score >= AMBIGUITY_MIN_SCORE_THRESHOLD
-                && top.score.abs_diff(s.score) <= AMBIGUITY_SCORE_GAP_THRESHOLD
-        })
-        .map(|(idx, _)| *idx)
-        .collect::<Vec<_>>();
-    if cluster.len() >= 2 {
-        cluster
-    } else {
-        Vec::new()
-    }
-}
-
-fn evidence_quality(run: &Run) -> EvidenceQuality {
-    let requests = request_status(run);
-    let queues = family_status(run.queues.is_empty(), run.truncation.dropped_queues);
-    let stages = family_status(run.stages.is_empty(), run.truncation.dropped_stages);
-    let runtime_snapshots = runtime_status(run);
-    let inflight_snapshots = family_status(
-        run.inflight.is_empty(),
-        run.truncation.dropped_inflight_snapshots,
-    );
-    let limitations = evidence_limitations(run, queues, stages, runtime_snapshots);
-    let non_request_truncated = matches!(queues, SignalCoverageStatus::Truncated)
-        || matches!(stages, SignalCoverageStatus::Truncated)
-        || matches!(runtime_snapshots, SignalCoverageStatus::Truncated)
-        || matches!(inflight_snapshots, SignalCoverageStatus::Truncated);
-    let explanatory_present =
-        !run.queues.is_empty() || !run.stages.is_empty() || !run.runtime_snapshots.is_empty();
-    let quality = if run.requests.is_empty()
-        || run.requests.len() < LOW_COMPLETED_REQUEST_THRESHOLD
-        || run.truncation.dropped_requests > 0
-        || !explanatory_present
-    {
-        EvidenceQualityLevel::Weak
-    } else if non_request_truncated
-        || (run.queues.is_empty() && run.stages.is_empty())
-        || runtime_snapshots == SignalCoverageStatus::Partial
-    {
-        EvidenceQualityLevel::Partial
-    } else {
-        // Runtime snapshots are optional; when queue/stage evidence is otherwise strong,
-        // missing runtime is represented as a limitation, not an automatic downgrade.
-        EvidenceQualityLevel::Strong
-    };
-
-    EvidenceQuality {
-        request_count: run.requests.len(),
-        queue_event_count: run.queues.len(),
-        stage_event_count: run.stages.len(),
-        runtime_snapshot_count: run.runtime_snapshots.len(),
-        inflight_snapshot_count: run.inflight.len(),
-        requests,
-        queues,
-        stages,
-        runtime_snapshots,
-        inflight_snapshots,
-        truncated: run.truncation.is_truncated() || run.truncation.limits_hit,
-        dropped_requests: run.truncation.dropped_requests,
-        dropped_stages: run.truncation.dropped_stages,
-        dropped_queues: run.truncation.dropped_queues,
-        dropped_inflight_snapshots: run.truncation.dropped_inflight_snapshots,
-        dropped_runtime_snapshots: run.truncation.dropped_runtime_snapshots,
-        quality,
-        limitations,
-    }
-}
-
-fn request_status(run: &Run) -> SignalCoverageStatus {
-    if run.requests.is_empty() {
-        SignalCoverageStatus::Missing
-    } else if run.truncation.dropped_requests > 0 {
-        SignalCoverageStatus::Truncated
-    } else if run.requests.len() < LOW_COMPLETED_REQUEST_THRESHOLD {
-        SignalCoverageStatus::Partial
-    } else {
-        SignalCoverageStatus::Present
-    }
-}
-
-fn family_status(is_empty: bool, dropped: u64) -> SignalCoverageStatus {
-    if dropped > 0 {
-        SignalCoverageStatus::Truncated
-    } else if is_empty {
-        SignalCoverageStatus::Missing
-    } else {
-        SignalCoverageStatus::Present
-    }
-}
-
-fn runtime_status(run: &Run) -> SignalCoverageStatus {
-    if run.truncation.dropped_runtime_snapshots > 0 {
-        SignalCoverageStatus::Truncated
-    } else if run.runtime_snapshots.is_empty() {
-        SignalCoverageStatus::Missing
-    } else if run
-        .runtime_snapshots
-        .iter()
-        .all(|snapshot| snapshot.blocking_queue_depth.is_none())
-        || run
-            .runtime_snapshots
-            .iter()
-            .all(|snapshot| snapshot.local_queue_depth.is_none())
-        || run
-            .runtime_snapshots
-            .iter()
-            .all(|snapshot| snapshot.global_queue_depth.is_none())
-    {
-        SignalCoverageStatus::Partial
-    } else {
-        SignalCoverageStatus::Present
-    }
-}
-
-fn evidence_limitations(
-    run: &Run,
-    queues: SignalCoverageStatus,
-    stages: SignalCoverageStatus,
-    runtime_snapshots: SignalCoverageStatus,
-) -> Vec<String> {
-    let mut limitations = Vec::new();
-    if run.requests.len() < LOW_COMPLETED_REQUEST_THRESHOLD {
-        limitations
-            .push("Low completed-request count can make suspect ranking unstable.".to_string());
-    }
-    if matches!(
-        queues,
-        SignalCoverageStatus::Missing | SignalCoverageStatus::Truncated
-    ) && matches!(
-        stages,
-        SignalCoverageStatus::Missing | SignalCoverageStatus::Truncated
-    ) {
-        limitations.push("Queue and stage instrumentation are both unavailable, limiting application vs downstream interpretation.".to_string());
-    }
-    if run.runtime_snapshots.is_empty() {
-        limitations.push("Runtime snapshots are missing, limiting executor and blocking-pressure interpretation.".to_string());
-    } else if runtime_snapshots == SignalCoverageStatus::Partial {
-        limitations.push("Runtime snapshots have missing queue-depth fields, limiting executor vs blocking differentiation.".to_string());
-    }
-    if run.truncation.is_truncated() || run.truncation.limits_hit {
-        limitations.push(
-            "Capture truncation dropped evidence and can reduce diagnosis completeness."
-                .to_string(),
-        );
-    }
-    limitations
-}
-
-fn truncation_warnings(run: &Run) -> Vec<String> {
-    let mut warnings = Vec::new();
-    if run.truncation.limits_hit || run.truncation.is_truncated() {
-        warnings.push(
-            "Capture limits were hit during this run; dropped evidence can reduce diagnosis completeness and confidence."
-                .to_string(),
-        );
-    }
-    if run.truncation.dropped_requests > 0 {
-        warnings.push(format!(
-            "Capture truncated requests: dropped {} request events after reaching the configured max_requests limit. This dropped evidence can reduce diagnosis completeness and confidence.",
-            run.truncation.dropped_requests
-        ));
-    }
-    if run.truncation.dropped_stages > 0 {
-        warnings.push(format!(
-            "Capture truncated stages: dropped {} stage events after reaching the configured max_stages limit. This dropped evidence can reduce diagnosis completeness and confidence.",
-            run.truncation.dropped_stages
-        ));
-    }
-    if run.truncation.dropped_queues > 0 {
-        warnings.push(format!(
-            "Capture truncated queues: dropped {} queue events after reaching the configured max_queues limit. This dropped evidence can reduce diagnosis completeness and confidence.",
-            run.truncation.dropped_queues
-        ));
-    }
-    if run.truncation.dropped_inflight_snapshots > 0 {
-        warnings.push(format!(
-            "Capture truncated in-flight snapshots: dropped {} entries after reaching max_inflight_snapshots. This dropped evidence can reduce diagnosis completeness and confidence.",
-            run.truncation.dropped_inflight_snapshots
-        ));
-    }
-    if run.truncation.dropped_runtime_snapshots > 0 {
-        warnings.push(format!(
-            "Capture truncated runtime snapshots: dropped {} entries after reaching max_runtime_snapshots. This dropped evidence can reduce diagnosis completeness and confidence.",
-            run.truncation.dropped_runtime_snapshots
-        ));
-    }
-    warnings
-}
-
-fn clamp_score(value: u64) -> u8 {
-    u8::try_from(value.min(100)).unwrap_or(100)
-}
-
-fn nonzero_sample_count(values: &[u64]) -> usize {
-    values.iter().filter(|&&v| v > 0).count()
-}
-
-fn max_or_zero(values: &[u64]) -> u64 {
-    values.iter().copied().max().unwrap_or(0)
-}
-
-fn score_sample_quality(sample_count: usize) -> u8 {
-    if sample_count >= SAMPLE_QUALITY_HIGH_SAMPLE_COUNT {
-        8
-    } else if sample_count >= SAMPLE_QUALITY_MEDIUM_SAMPLE_COUNT {
-        5
-    } else if sample_count >= SAMPLE_QUALITY_LOW_SAMPLE_COUNT {
-        3
-    } else {
-        u8::from(sample_count >= SAMPLE_QUALITY_MIN_NONZERO_SAMPLE_COUNT)
-    }
-}
-
-fn score_from_permille(base: u64, permille: u64, scale: u64) -> u64 {
-    base + permille.min(1000) / scale
-}
-
-fn cap_unless_clean_evidence(score: u64, clean: bool, soft_cap: u8) -> u8 {
-    if clean {
-        clamp_score(score)
-    } else {
-        clamp_score(score.min(u64::from(soft_cap)))
-    }
-}
-
-fn queue_saturation_suspect(run: &Run, inflight_trend: Option<&InflightTrend>) -> Option<Suspect> {
-    let (queue_shares, _) = request_time_shares(run);
-    let p95_queue_share_permille = percentile(&queue_shares, 95, 100)?;
-    if p95_queue_share_permille < QUEUE_SHARE_TRIGGER_PERMILLE {
-        return None;
-    }
-    let queue_depths = run
-        .queues
-        .iter()
-        .filter_map(|q| q.depth_at_start)
-        .collect::<Vec<_>>();
-    let max_depth = max_or_zero(&queue_depths);
-    let growth_bonus = inflight_trend
-        .filter(|t| t.growth_delta > 0)
-        .map_or(0, |_| 5);
-    let depth_bonus = (max_depth.min(40) * 2) / 3;
-    let base = score_from_permille(22, p95_queue_share_permille, 14);
-    let clean_extreme = p95_queue_share_permille >= 985
-        && max_depth >= 12
-        && queue_shares.len() >= 20
-        && inflight_trend.is_some_and(|t| t.growth_delta > 0);
-    let score = cap_unless_clean_evidence(
-        base + depth_bonus + growth_bonus + u64::from(score_sample_quality(queue_shares.len())),
-        clean_extreme,
-        95,
-    );
-    let mut evidence = vec![format!(
-        "Queue wait at p95 consumes {}.{}% of request time.",
-        p95_queue_share_permille / 10,
-        p95_queue_share_permille % 10
-    )];
-    if max_depth > 0 {
-        evidence.push(format!("Observed queue depth sample up to {max_depth}."));
-    }
-    if let Some(trend) = inflight_trend.filter(|trend| trend.growth_delta > 0) {
-        evidence.push(format!(
-            "In-flight gauge '{}' grew by {} over the run window (p95={}, peak={}).",
-            trend.gauge, trend.growth_delta, trend.p95_count, trend.peak_count
-        ));
-    }
-    Some(Suspect::new(
-        DiagnosisKind::ApplicationQueueSaturation,
-        score,
-        evidence,
-        vec![
-            "Inspect queue admission limits and producer burst patterns.".to_string(),
-            "Compare queue wait distribution before and after increasing worker parallelism."
-                .to_string(),
-        ],
-    ))
-}
-
-#[derive(Clone, Copy)]
-struct BlockingSignal {
-    p95: u64,
-    peak: u64,
-    nonzero: usize,
-    samples: usize,
-    nz_share_permille: u64,
-}
-
-fn blocking_signal(run: &Run) -> Option<BlockingSignal> {
-    let depths = runtime_metric_series(&run.runtime_snapshots, |s| s.blocking_queue_depth);
-    let p95 = percentile(&depths, 95, 100)?;
-    let nonzero = nonzero_sample_count(&depths);
-    if p95 == 0 && nonzero < 2 {
-        return None;
-    }
-    let peak = max_or_zero(&depths);
-    let nz_share_permille = if depths.is_empty() {
-        0
-    } else {
-        nonzero as u64 * 1000 / depths.len() as u64
-    };
-    Some(BlockingSignal {
-        p95,
-        peak,
-        nonzero,
-        samples: depths.len(),
-        nz_share_permille,
-    })
-}
-
-fn strong_blocking_signal(signal: BlockingSignal) -> bool {
-    signal.p95 >= 12 && signal.peak >= 20 && signal.nz_share_permille >= 700 && signal.samples >= 30
-}
-
-fn stage_correlates_with_blocking_pool(stage: &str) -> bool {
-    let lower = stage.to_ascii_lowercase();
-    lower.contains("spawn_blocking")
-        || lower.contains("blocking_path")
-        || lower.contains("blocking")
-}
-
-fn blocking_pressure_suspect(run: &Run) -> Option<Suspect> {
-    let signal = blocking_signal(run)?;
-    let clean_extreme = signal.p95 >= 16 && signal.peak >= 24 && signal.nz_share_permille >= 900;
-    let score = cap_unless_clean_evidence(
-        32 + signal.p95.min(24)
-            + (signal.peak.min(24) / 2)
-            + (signal.nz_share_permille / 80)
-            + u64::from(score_sample_quality(signal.samples)),
-        clean_extreme,
-        94,
-    );
-    Some(Suspect::new(
-        DiagnosisKind::BlockingPoolPressure,
-        score,
-        vec![format!(
-            "Blocking queue depth p95 is {}, peak is {}, with {}/{} nonzero samples.",
-            signal.p95, signal.peak, signal.nonzero, signal.samples
-        )],
-        vec![
-            "Audit blocking sections and move avoidable synchronous work out of hot paths."
-                .to_string(),
-            "Inspect spawn_blocking callsites for long-running CPU or I/O work.".to_string(),
-        ],
-    ))
-}
-
-fn executor_pressure_suspect(run: &Run, inflight_trend: Option<&InflightTrend>) -> Option<Suspect> {
-    let global = runtime_metric_series(&run.runtime_snapshots, |s| s.global_queue_depth);
-    let p95_global = percentile(&global, 95, 100)?;
-    if p95_global == 0 {
-        return None;
-    }
-    let local = runtime_metric_series(&run.runtime_snapshots, |s| s.local_queue_depth);
-    let alive = runtime_metric_series(&run.runtime_snapshots, |s| s.alive_tasks);
-    let growth_bonus = inflight_trend
-        .filter(|t| t.growth_delta > 0)
-        .map_or(0, |_| 4);
-    let clean_extreme = p95_global >= 140 && global.len() >= 30;
-    let score = cap_unless_clean_evidence(
-        34 + (p95_global.min(150) / 4)
-            + (percentile(&local, 95, 100).unwrap_or(0).min(60) / 6)
-            + (percentile(&alive, 95, 100).unwrap_or(0).min(400) / 40)
-            + growth_bonus
-            + u64::from(score_sample_quality(global.len())),
-        clean_extreme,
-        94,
-    );
-    let mut evidence = vec![format!(
-        "Runtime global queue depth p95 is {p95_global}, suggesting scheduler contention."
-    )];
-    if let Some(lp95) = percentile(&local, 95, 100) {
-        evidence.push(format!("Runtime local queue depth p95 is {lp95}."));
-    }
-    if let Some(ap95) = percentile(&alive, 95, 100) {
-        evidence.push(format!("Runtime alive_tasks p95 is {ap95}."));
-    }
-    Some(Suspect::new(
-        DiagnosisKind::ExecutorPressureSuspected,
-        score,
-        evidence,
-        vec![
-            "Check for long polls without yielding and uneven task fan-out.".to_string(),
-            "Compare with per-stage timings to isolate overloaded async stages.".to_string(),
-        ],
-    ))
-}
-
-#[derive(Clone)]
-struct StageCandidate {
-    stage: String,
-    samples: usize,
-    p95: u64,
-    cumulative: u64,
-    cum_share: u64,
-    tail_share: u64,
-    score: u8,
-}
-
-fn downstream_stage_candidates(run: &Run, p95_req: u64, total_req: u64) -> Vec<StageCandidate> {
-    let tail_ids: std::collections::HashMap<&str, u64> = run
-        .requests
-        .iter()
-        .filter(|r| r.latency_us >= p95_req)
-        .map(|r| (r.request_id.as_str(), r.latency_us))
-        .collect();
-    let tail_total = tail_ids.values().copied().fold(0_u64, u64::saturating_add);
-    let mut by: BTreeMap<&str, Vec<&tailtriage_core::StageEvent>> = BTreeMap::new();
-    for st in &run.stages {
-        by.entry(st.stage.as_str()).or_default().push(st);
-    }
-    let mut cands = Vec::new();
-    for (name, ss) in by {
-        if ss.len() < DOWNSTREAM_MIN_STAGE_SAMPLES {
-            continue;
-        }
-        let lats = ss.iter().map(|s| s.latency_us).collect::<Vec<_>>();
-        let cum = lats.iter().copied().fold(0_u64, u64::saturating_add);
-        let p95 = percentile(&lats, 95, 100).unwrap_or(0);
-        let cum_share = cum.saturating_mul(1000).checked_div(total_req).unwrap_or(0);
-        let tail_stage = ss
-            .iter()
-            .filter_map(|s| tail_ids.get(s.request_id.as_str()).map(|_| s.latency_us))
-            .fold(0_u64, u64::saturating_add);
-        let tail_share = if tail_total == 0 {
-            0
-        } else {
-            tail_stage
-                .saturating_mul(1000)
-                .checked_div(tail_total)
-                .unwrap_or(0)
-        };
-        let clean_extreme = tail_share >= 960 && cum_share >= 920 && ss.len() >= 20;
-        let score = cap_unless_clean_evidence(
-            score_from_permille(24, tail_share, 11)
-                + (cum_share / 35)
-                + u64::from(score_sample_quality(ss.len())),
-            clean_extreme,
-            95,
-        );
-        cands.push(StageCandidate {
-            stage: name.to_string(),
-            samples: ss.len(),
-            p95,
-            cumulative: cum,
-            cum_share,
-            tail_share,
-            score,
-        });
-    }
-    cands
-}
-
-fn downstream_stage_suspect(run: &Run) -> Option<Suspect> {
-    let p95_req = percentile(
-        &run.requests
-            .iter()
-            .map(|r| r.latency_us)
-            .collect::<Vec<_>>(),
-        95,
-        100,
-    )?;
-    let total_req = run
-        .requests
-        .iter()
-        .map(|r| r.latency_us)
-        .fold(0_u64, u64::saturating_add);
-    let blocking = blocking_signal(run);
-    let blocking_score = blocking.map(|signal| {
-        let clean_extreme =
-            signal.p95 >= 16 && signal.peak >= 24 && signal.nz_share_permille >= 900;
-        cap_unless_clean_evidence(
-            32 + signal.p95.min(24)
-                + (signal.peak.min(24) / 2)
-                + (signal.nz_share_permille / 80)
-                + u64::from(score_sample_quality(signal.samples)),
-            clean_extreme,
-            94,
-        )
-    });
-    let best = downstream_stage_candidates(run, p95_req, total_req)
-        .into_iter()
-        .max_by(|a, b| {
-            a.score
-                .cmp(&b.score)
-                .then_with(|| a.tail_share.cmp(&b.tail_share))
-                .then_with(|| a.cum_share.cmp(&b.cum_share))
-                .then_with(|| b.stage.cmp(&a.stage))
-        })?;
-    let mut downstream_score = best.score;
-    let mut correlation_evidence: Option<String> = None;
-    if stage_correlates_with_blocking_pool(&best.stage)
-        && blocking.is_some_and(strong_blocking_signal)
-        && blocking_score.is_some()
-    {
-        let cap = blocking_score.unwrap_or(downstream_score).saturating_sub(2);
-        downstream_score = downstream_score.min(cap);
-        correlation_evidence = Some(format!(
-            "Stage '{}' looks blocking-correlated; strong runtime blocking-queue evidence keeps blocking_pool_pressure prioritized.",
-            best.stage
-        ));
-    }
-    let mut evidence = vec![
-        format!(
-            "Stage '{}' has p95 latency {} us across {} samples.",
-            best.stage, best.p95, best.samples
-        ),
-        format!(
-            "Stage '{}' cumulative latency is {} us ({} permille of request latency).",
-            best.stage, best.cumulative, best.cum_share
-        ),
-        format!(
-            "Stage '{}' contributes {} permille of tail request latency.",
-            best.stage, best.tail_share
-        ),
-    ];
-    if let Some(extra) = correlation_evidence {
-        evidence.push(extra);
-    }
-    Some(Suspect::new(
-        DiagnosisKind::DownstreamStageDominates,
-        downstream_score,
-        evidence,
-        vec![
-            format!(
-                "Inspect downstream dependency behind stage '{}'.",
-                best.stage
-            ),
-            "Collect downstream service timings and retry behavior during tail windows."
-                .to_string(),
-            "Review downstream SLO/error budget and align retry budget/backoff with it."
-                .to_string(),
-        ],
-    ))
-}
-
 fn ambiguity_warning(suspects: &[Suspect]) -> Option<String> {
     let mut ranked = suspects
         .iter()
@@ -1645,152 +675,6 @@ fn percentile_sorted_u64(values: &[u64], numerator: usize, denominator: usize) -
         .div_ceil(denominator)
         .min(max_index);
     values.get(index).copied()
-}
-
-fn fmt_opt_u64(value: Option<u64>) -> String {
-    match value {
-        Some(value) => value.to_string(),
-        None => "n/a".to_string(),
-    }
-}
-
-fn fmt_percent_permille(value: Option<u64>) -> String {
-    match value {
-        Some(value) => format!("{}.{:01}%", value / 10, value % 10),
-        None => "n/a".to_string(),
-    }
-}
-
-fn fmt_confidence(confidence: Confidence) -> &'static str {
-    match confidence {
-        Confidence::Low => "low",
-        Confidence::Medium => "medium",
-        Confidence::High => "high",
-    }
-}
-
-#[must_use]
-/// Renders a compact text triage summary from a [`Report`].
-///
-/// The rendered output is guidance for follow-up checks, not proof of root cause.
-pub fn render_text(report: &Report) -> String {
-    let mut lines = vec![
-        "tailtriage diagnosis".to_string(),
-        format!("Requests analyzed: {}", report.request_count),
-        format!(
-            "Latency (us): p50 {}, p95 {}, p99 {}",
-            fmt_opt_u64(report.p50_latency_us),
-            fmt_opt_u64(report.p95_latency_us),
-            fmt_opt_u64(report.p99_latency_us),
-        ),
-        format!(
-            "Request time at p95: queue {}, non-queue service {}",
-            fmt_percent_permille(report.p95_queue_share_permille),
-            fmt_percent_permille(report.p95_service_share_permille),
-        ),
-    ];
-
-    match &report.inflight_trend {
-        Some(trend) => {
-            lines.push(format!(
-                "Inflight trend: gauge '{}', samples {}, peak {}, p95 {}, net growth {:+}",
-                trend.gauge,
-                trend.sample_count,
-                trend.peak_count,
-                trend.p95_count,
-                trend.growth_delta,
-            ));
-        }
-        None => {
-            lines.push("Inflight trend: none".to_string());
-        }
-    }
-
-    lines.push(format!(
-        "Primary suspect: {} ({} confidence, score {})",
-        report.primary_suspect.kind.as_str(),
-        fmt_confidence(report.primary_suspect.confidence),
-        report.primary_suspect.score,
-    ));
-    lines.push(format!(
-        "Evidence quality: {}{}",
-        match report.evidence_quality.quality {
-            EvidenceQualityLevel::Strong => "strong",
-            EvidenceQualityLevel::Partial => "partial",
-            EvidenceQualityLevel::Weak => "weak",
-        },
-        report
-            .evidence_quality
-            .limitations
-            .first()
-            .map_or_else(String::new, |l| format!(" ({l})"))
-    ));
-
-    if !report.warnings.is_empty() {
-        lines.push("Warnings:".to_string());
-        for warning in &report.warnings {
-            lines.push(format!("- {warning}"));
-        }
-    }
-
-    if !report.primary_suspect.evidence.is_empty() {
-        lines.push("Evidence:".to_string());
-        for evidence in &report.primary_suspect.evidence {
-            lines.push(format!("- {evidence}"));
-        }
-    }
-
-    if !report.primary_suspect.next_checks.is_empty() {
-        lines.push("Next checks:".to_string());
-        for next_check in &report.primary_suspect.next_checks {
-            lines.push(format!("- {next_check}"));
-        }
-    }
-
-    if !report.secondary_suspects.is_empty() {
-        lines.push("Secondary suspects:".to_string());
-        for suspect in &report.secondary_suspects {
-            lines.push(format!(
-                "- {} ({} confidence, score {})",
-                suspect.kind.as_str(),
-                fmt_confidence(suspect.confidence),
-                suspect.score,
-            ));
-        }
-    }
-    if !report.route_breakdowns.is_empty() {
-        lines.push("Route breakdowns:".to_string());
-        for route in &report.route_breakdowns {
-            lines.push(format!(
-                "- {}: requests {}, p95 {}us, suspect {} ({} confidence)",
-                route.route,
-                route.request_count,
-                fmt_opt_u64(route.p95_latency_us),
-                route.primary_suspect.kind.as_str(),
-                fmt_confidence(route.primary_suspect.confidence),
-            ));
-        }
-    }
-    append_temporal_segment_text(&mut lines, &report.temporal_segments);
-
-    lines.join("\n")
-}
-
-fn append_temporal_segment_text(lines: &mut Vec<String>, segments: &[TemporalSegment]) {
-    if segments.is_empty() {
-        return;
-    }
-    lines.push("Temporal segments:".to_string());
-    for seg in segments {
-        lines.push(format!(
-            "- {}: requests {}, p95 {}us, suspect {} ({} confidence)",
-            seg.name,
-            seg.request_count,
-            fmt_opt_u64(seg.p95_latency_us),
-            seg.primary_suspect.kind.as_str(),
-            fmt_confidence(seg.primary_suspect.confidence),
-        ));
-    }
 }
 
 #[cfg(test)]
